+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fixed line of YAML file for master deployments via docker-compose, so that image of excel export service is pulled correctly [#223](https://github.com/openkfw/TruBudget/issues/223)
-
+- Backup/restore works again. [#237](https://github.com/openkfw/TruBudget/issues/237)
 
 ## [1.0.0-beta.9] - 2019-04-23